% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http_server_clients.R
\name{http_server_clients}
\alias{http_server_clients}
\alias{HttpClientsNames}
\title{Get list of all names of clients that already connected to the http server}
\usage{
http_server_clients()

HttpClientsNames()
}
\value{
A character vector with the name of currently connected clients.
}
\description{
Get list of all names of clients that already connected to the http server
}
\examples{
library(svHttp)
http_server_clients()
}
\seealso{
\code{\link[=start_http_server]{start_http_server()}} for a complete example.
}
\concept{Interprocess communication}
\keyword{IO}
