% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http_server_port.R
\name{http_server_port}
\alias{http_server_port}
\alias{HttpServerPort}
\title{Get or change the port of the HTTP server}
\usage{
http_server_port(port)

HttpServerPort(port)
}
\arguments{
\item{port}{port on which the server should run (both help and SciViews). By
default, it is port 8888. Note that this server runs only locally and can
only serve requests from 127.0.0.1 (because communication is not crypted).}
}
\value{
A number with the port of the HTTP server.
}
\description{
Get or change the port of the HTTP server
}
\examples{
http_server_port()
}
\seealso{
\code{\link[=start_http_server]{start_http_server()}} for a complete example.
}
\concept{Interprocess communication}
\keyword{IO}
