% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_http_server.R
\name{par_http_server}
\alias{par_http_server}
\alias{parHttp}
\title{Get or change http server options}
\usage{
par_http_server(client, ...)

parHttp(client, ...)
}
\arguments{
\item{client}{the name of one client. A client that does not identify
itself is named \code{default}.}

\item{...}{named arguments specifying options to set or change.}
}
\value{
An environment that contains the whole configuration is returned
invisibly.
}
\description{
Get or change http server options
}
\note{
Possible named arguments (with their default values) are: \code{prompt = ":> "}
for the server prompt, \code{continue = ":+ "} for the continuation prompt when
multiline instructions are send, \code{code = ""} for current partial code in
multiline mode, \code{last = ""} for a string to add at the end of each
evaluation, \code{echo = FALSE} to echo commands at the R console or terminal,
\code{multiline = TRUE} to allow multiline mode, \code{bare = TRUE} a bare mode that
inactivates all other options (the server is always started in bare mode).
}
\seealso{
\code{\link[=start_http_server]{start_http_server()}} for a complete example.
}
\concept{Interprocess communication}
\keyword{IO}
