% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_http_server.R
\name{stop_http_server}
\alias{stop_http_server}
\alias{stopHttpSever}
\title{Stop the SciViews and R HTTP server and eliminate all tracks}
\usage{
stop_http_server(remove.clients = FALSE)

stopHttpSever(remove.clients = FALSE)
}
\arguments{
\item{remove.clients}{do we remove also persistent data for the clients,
\code{FALSE} by default.}
}
\value{
A number with the port of the HTTP server.
}
\description{
Stop the SciViews and R HTTP server and eliminate all tracks
}
\seealso{
\code{\link[=start_http_server]{start_http_server()}} for a complete example.
}
\concept{Interprocess communication}
\keyword{IO}
