% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svHttp-package.R
\docType{package}
\name{svHttp-package}
\alias{svHttp}
\alias{svHttp-package}
\title{svHttp: 'SciViews' - HTTP Server}
\description{
A simple HTTP server allows to connect GUI clients to R.
}
\details{
A SciViews R server using HTTP R help server and JSONP for communication.

See \code{\link[=start_http_server]{start_http_server()}} for further implementation details.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SciViews/svHttp}
  \item \url{https://www.sciviews.org/svHttp/}
  \item Report bugs at \url{https://github.com/SciViews/svHttp/issues}
}

}
\author{
\strong{Maintainer}: Philippe Grosjean \email{phgrosjean@sciviews.org} (\href{https://orcid.org/0000-0002-2694-9471}{ORCID})

}
\keyword{internal}
