\name{svHttp-package}
\alias{svHttp-package}
\alias{svHttp}
\docType{package}

\title{ SciViews GUI API - R HTTP server }
\description{
  A HTTP server based on default R help HTTP server introduced in R 2.10.0, used
  for local intercommunication with other programs, like SciViews Komodo (see
  the svKomodo package). It is mostly compatible with the socket server in
  svSocket, but does not require Tcl/Tk and the tcltk R package to run.
}
\details{
  \tabular{ll}{
    Package: \tab svHttp\cr
    Type: \tab Package\cr
    Version: \tab 0.9-54\cr
    Date: \tab 2012-04-02\cr
    License: \tab GPL 2 or above, at your convenience\cr
  }
  % TODO: add description of main functions here. Also add examples
}

\author{
  Philippe Grosjean

  Maintainer: Ph. Grosjean <phgrosjean@sciviews.org>
}

\keyword{ package }
\keyword{ utilities }

\concept{ GUI API HTTP server }
